Move provider enums to MediaEmbed\Enum namespace
ProviderStatus and ProviderCategory move from MediaEmbed\Provider to a dedicated
MediaEmbed\Enum namespace (src/Enum/) for clarity. Update imports, docblocks and
UPGRADE references accordingly.
